requirements.txt
asgiref==3.2.7
boto3==1.12.39
botocore==1.15.39
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
numpy==1.18.2
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.7.5
psycopg2-binary==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
